Add a 5 ms TimeDelay on every WarpPrint
When the processor is in RUN mode the RTT isn't fast enough at consuming
the buffer before the processor writes again. What happens is lines or
characters are missing when in RUN mode.

This commit introduces a short 5ms delay at the end of each warpPrint.
The delay is short enough to not be noticeable when printing but long
enough so that everything prints correctly.

It's possible that I'm missing something here like the ability to speed
up the RTTClient or to increase the buffer length.

Addresses #101.