IS25xP driver (mostly) for Glaux (#102)
* Set PTA9 as nCS for IS25xP when on FRDMKL03

Addresses #101.

* Enhance IS25xP device driver

- Add buffer overflow check for spiTransaction.

Expand on the functionality based on the spiTransaction function to:
- Add readMemory function for contiguous Normal Read.
- Add programPage function for programming up to 256 bytes of memory on
  the same flash page. (WiP)
- Enhance documentation of header file.

I haven't been able to make the programPage function to work even when
sending WREN (Write Enable) instruction before the write operations.

The SPI communication fails when the processor is in the VLPR
mode.

Addresses #101.

* Add a 5 ms TimeDelay on every WarpPrint

When the processor is in RUN mode the RTT isn't fast enough at consuming
the buffer before the processor writes again. What happens is lines or
characters are missing when in RUN mode.

This commit introduces a short 5ms delay at the end of each warpPrint.
The delay is short enough to not be noticeable when printing but long
enough so that everything prints correctly.

It's possible that I'm missing something here like the ability to speed
up the RTTClient or to increase the buffer length.

Addresses #101.

* Add menu for flash device

Add a main menu item for input 'F' that brings up a menu specific for
the flash device. It's a work in progress but it currently allows to
dump informational registers and read memory chunks.

Write is not working properly even though the WREN command is sent.

Addresses #101.

* Add erase helpers

Addresses #101.

* Enhance and rearrange code in flash menu

Addresses #101.